Move useful matchers to helper module
